+++ src/Animation.h
@@ -3,23 +3,23 @@
 #include <MyKnob.h>
 #include "FastLED.h"
 #include "config.h"
+#include "BaseControllers.h"
 
-class Animation {
+class Animation : public BaseController {
  protected:
   CRGB *leds;
   bool configChangeFlag = false;
 
-  virtual void update() = 0;
+  virtual bool updateAnimation(bool justActivated) = 0;
+
+  virtual void loop(bool justActivated) override {
+    configChangeFlag = false;
+    if (updateAnimation(justActivated))
+      FastLED.show();
+  };
 
  public:
   Animation(CRGB leds_[]) : leds(leds_) {};
-
-  virtual void wakeUp() = 0;
-
-  virtual void loop() {
-    configChangeFlag = false;
-    update();
-  };
 
   virtual bool hasConfigChanged() { return configChangeFlag; }
 };
